Introduce ISA system
This is shared code between the VM and the compiler.
In the beginning we don't bother going all the way
to encode instructions and constants into bytecode.
Instead we're going to deal with a chunk which
is kind of an already decoded piece of data, which
holds bytecode instructions and constants as
common lisp values already