Generate instructions (#9)
This adds the basic machinary to generate bytecode. I opted to go for a mechanism that uses the description of an instruction set to drive code generation which also serves as a way to document and check instructions during generation.

I also added a disassembler which uses the same information to disassemble instructions in a chunk.
I was careful to make use of proper types and more elaborate objects during generation of chunks,
but the actual representation of an instruction is not typesafe to make it fast to read and easy to decode.