compiler: parse assignment
This took a bigger change in how we parse.
In order to enable backtracking we introduced a parse buffer,
which is a vector of all tokens in the input stream.
This way backtracking becomes trivial and in exchange
our parse can be encoded more closely to the grammar.

We will need the full token stream in the future anyway
to send it to the introspection view.